requirements.txt
torch
torchvision
pyyaml
easydict